// Web/3-weather/src/types/Weather.tsx
export default interface Weather {
  header: {
    city: string;
    country: string;
    temperature: number;
    unit: string;
  };
  stats: {
    iconUrl: string;
    title: string;
    value: string;
    unit: string;
  }[];
  forecast: {
    day: string;
    iconUrl: string;
    highTemp: string;
    lowTemp: string;
    precipitation: string;
  }[];
}